# internal/pkg/transport/http/openapi.yaml
openapi: 3.0.3
info:
  title: Сервис агрегации онлайн-подписок
  version: 1.0.0
  description: API для управления пользовательскими онлайн-подписками и расчета их суммарной стоимости за период.
servers:
  - url: /
tags:
  - name: Подписки
    description: Операции управления подписками
  - name: Отчеты
    description: Операции агрегации стоимости подписок
paths:
  /api/v1/subscriptions:
    get:
      tags:
        - Подписки
      operationId: ListSubscriptions
      summary: Получить список подписок
      description: Возвращает список подписок с возможностью фильтрации по пользователю и названию сервиса.
      parameters:
        - in: query
          name: user_id
          description: Идентификатор пользователя в формате UUID.
          schema:
            type: string
            format: uuid
        - in: query
          name: service_name
          description: Название сервиса подписки.
          schema:
            type: string
            minLength: 1
            maxLength: 255
        - in: query
          name: limit
          description: Максимальное количество записей в ответе.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: offset
          description: Смещение относительно начала списка.
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Список подписок.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - Подписки
      operationId: CreateSubscription
      summary: Создать подписку
      description: Создает новую запись о подписке пользователя.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionRequest'
      responses:
        '201':
          description: Подписка успешно создана.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/subscriptions/{subscription_id}:
    parameters:
      - in: path
        name: subscription_id
        required: true
        description: Идентификатор записи о подписке.
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Подписки
      operationId: GetSubscription
      summary: Получить подписку по идентификатору
      description: Возвращает запись о подписке по ее идентификатору.
      responses:
        '200':
          description: Подписка найдена.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags:
        - Подписки
      operationId: UpdateSubscription
      summary: Обновить подписку
      description: Полностью обновляет существующую запись о подписке.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSubscriptionRequest'
      responses:
        '200':
          description: Подписка обновлена.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags:
        - Подписки
      operationId: DeleteSubscription
      summary: Удалить подписку
      description: Удаляет запись о подписке по ее идентификатору.
      responses:
        '204':
          description: Подписка удалена.
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/reports/total-cost:
    get:
      tags:
        - Отчеты
      operationId: CalculateTotalCost
      summary: Посчитать суммарную стоимость подписок
      description: Возвращает суммарную стоимость подписок за выбранный период с фильтрацией по пользователю и названию сервиса.
      parameters:
        - in: query
          name: start_period
          required: true
          description: Начало периода в формате MM-YYYY.
          schema:
            $ref: '#/components/schemas/MonthYear'
        - in: query
          name: end_period
          required: true
          description: Конец периода в формате MM-YYYY.
          schema:
            $ref: '#/components/schemas/MonthYear'
        - in: query
          name: user_id
          description: Идентификатор пользователя в формате UUID.
          schema:
            type: string
            format: uuid
        - in: query
          name: service_name
          description: Название сервиса подписки.
          schema:
            type: string
            minLength: 1
            maxLength: 255
      responses:
        '200':
          description: Суммарная стоимость подписок за период.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TotalCostResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  responses:
    BadRequest:
      description: Некорректные параметры запроса.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Запись не найдена.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Конфликт данных.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Внутренняя ошибка сервера.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    MonthYear:
      type: string
      pattern: '^(0[1-9]|1[0-2])-[0-9]{4}$'
      example: '07-2025'
      description: Месяц и год в формате MM-YYYY.
    CreateSubscriptionRequest:
      type: object
      required:
        - service_name
        - price
        - user_id
        - start_date
      properties:
        service_name:
          type: string
          minLength: 1
          maxLength: 255
          description: Название сервиса, предоставляющего подписку.
          example: Yandex Plus
        price:
          type: integer
          minimum: 1
          description: Стоимость месячной подписки в рублях.
          example: 400
        user_id:
          type: string
          format: uuid
          description: Идентификатор пользователя.
          example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        start_date:
          $ref: '#/components/schemas/MonthYear'
        end_date:
          allOf:
            - $ref: '#/components/schemas/MonthYear'
          nullable: true
          description: Дата окончания подписки в формате MM-YYYY.
    UpdateSubscriptionRequest:
      type: object
      required:
        - service_name
        - price
        - user_id
        - start_date
      properties:
        service_name:
          type: string
          minLength: 1
          maxLength: 255
          description: Название сервиса, предоставляющего подписку.
        price:
          type: integer
          minimum: 1
          description: Стоимость месячной подписки в рублях.
        user_id:
          type: string
          format: uuid
          description: Идентификатор пользователя.
        start_date:
          $ref: '#/components/schemas/MonthYear'
        end_date:
          allOf:
            - $ref: '#/components/schemas/MonthYear'
          nullable: true
          description: Дата окончания подписки в формате MM-YYYY.
    SubscriptionResponse:
      type: object
      required:
        - id
        - service_name
        - price
        - user_id
        - start_date
      properties:
        id:
          type: string
          format: uuid
          description: Идентификатор записи.
        service_name:
          type: string
          description: Название сервиса, предоставляющего подписку.
        price:
          type: integer
          description: Стоимость месячной подписки в рублях.
        user_id:
          type: string
          format: uuid
          description: Идентификатор пользователя.
        start_date:
          $ref: '#/components/schemas/MonthYear'
        end_date:
          allOf:
            - $ref: '#/components/schemas/MonthYear'
          nullable: true
          description: Дата окончания подписки в формате MM-YYYY.
    SubscriptionListResponse:
      type: object
      required:
        - items
        - total
        - limit
        - offset
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SubscriptionResponse'
        total:
          type: integer
          format: int64
          description: Общее количество записей, удовлетворяющих фильтрам.
        limit:
          type: integer
          description: Лимит выдачи.
        offset:
          type: integer
          description: Смещение выдачи.
    TotalCostResponse:
      type: object
      required:
        - total_cost
        - currency
        - start_period
        - end_period
      properties:
        total_cost:
          type: integer
          description: Суммарная стоимость подписок за выбранный период в рублях.
          example: 2400
        currency:
          type: string
          description: Валюта расчета.
          example: RUB
        start_period:
          $ref: '#/components/schemas/MonthYear'
        end_period:
          $ref: '#/components/schemas/MonthYear'
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Текст ошибки.